+++ libs/ng-observable/lib/ng-observable.ts
@@ -11,44 +11,26 @@ import {
 } from "@angular/core"
 import { InvokeSubject } from "./invoke-subject"
 import { mixins } from "./internals/apply-mixins"
-import { createMask } from "./internals/create-mask"
+import { UseHooks } from "./internals/use-hooks"
 
-export function UseHooks(...flags: number[]) {
-    const features = createMask(...flags)
-    return function(target: any) {
-        if (features & 1) {
-            target.prototype.ngOnChanges = new InvokeSubject(target.prototype.ngOnChanges)
-        }
-        if (features & 2) {
-            target.prototype.ngOnInit = new InvokeSubject(target.prototype.ngOnInit)
-        }
-        if (features & 4) {
-            target.prototype.ngDoCheck = new InvokeSubject(target.prototype.ngDoCheck)
-        }
-        if (features & 8) {
-            target.prototype.ngAfterContentInit = new InvokeSubject(
-                target.prototype.ngAfterContentInit,
-            )
-        }
-        if (features & 16) {
-            target.prototype.ngAfterContentChecked = new InvokeSubject(
-                target.prototype.ngAfterContentChecked,
-            )
-        }
-        if (features & 32) {
-            target.prototype.ngAfterViewInit = new InvokeSubject(target.prototype.ngAfterViewInit)
-        }
-        if (features & 64) {
-            target.prototype.ngAfterViewChecked = new InvokeSubject(
-                target.prototype.ngAfterViewChecked,
-            )
-        }
-        if (features & 128) {
-            target.prototype.ngOnDestroy = new InvokeSubject(target.prototype.ngOnDestroy)
-        }
-    }
-}
-
+/**
+ * Implements all of the builtin lifecycle hooks provided by Angular and makes them observable.
+ *
+ * @usageNotes
+ *
+ * ```ts
+ * @Component()
+ * export class MyComponent extends NgObservable {
+ *     constructor() {
+ *         ngOnInit(this).subscribe(() => {
+ *             console.log("ngOnInit")
+ *         })
+ *     }
+ * }
+ * ```
+ *
+ * @publicApi
+ */
 export abstract class NgObservable
     implements
         OnInit,
@@ -61,34 +43,60 @@ export abstract class NgObservable
         AfterViewChecked {
     static [mixins] = UseHooks(255)(NgObservable)
 
+    /**
+     * OnChanges lifecycle hook
+     *
+     * @param changes Simple changes can be strongly typed where input props are known
+     */
     public ngOnChanges(changes: TypedChanges<any>): void {
         ;(<InvokeSubject<any>>this.ngOnChanges).next([this, changes])
     }
 
+    /**
+     * OnInit lifecycle hook
+     */
     public ngOnInit(): void {
         ;(<InvokeSubject<any>>this.ngOnInit).next(this)
     }
 
+    /**
+     * ngDoCheck lifecycle hook
+     */
     public ngDoCheck() {
         ;(<InvokeSubject<any>>this.ngDoCheck).next(this)
     }
 
+    /**
+     * ngAfterContentInit lifecycle hook
+     */
     public ngAfterContentInit() {
         ;(<InvokeSubject<any>>this.ngAfterContentInit).next(this)
     }
 
+    /**
+     * ngAfterContentChecked lifecycle hook
+     */
     public ngAfterContentChecked() {
         ;(<InvokeSubject<any>>this.ngAfterContentChecked).next(this)
     }
 
+    /**
+     * ngAfterViewInit lifecycle hook
+     */
     public ngAfterViewInit() {
         ;(<InvokeSubject<any>>this.ngAfterViewInit).next(this)
     }
 
+    /**
+     * ngAfterViewChecked lifecycle hook
+     */
     public ngAfterViewChecked() {
         ;(<InvokeSubject<any>>this.ngAfterViewChecked).next(this)
     }
 
+    /**
+     * ngOnDestroy lifecycle hook
+     */
     public ngOnDestroy() {
         ;(<InvokeSubject<any>>this.ngOnDestroy).next(this)
     }