+++ libs/ng-observable/lib/stream.spec.ts
@@ -1,0 +1,7 @@
+import { Stream } from "./stream"
+
+describe("Sink", () => {
+    it("should create an instance", () => {
+        expect(new Stream()).toBeTruthy()
+    })
+})